Extract ConstraintExpression class into its own target (NFC).
It is usually a good idea to put separate classes into their own targets.
This makes merge conflicts less likely and can also mean that fewer tests will
be triggered (e.g. only the SymbolicTile tests when changing something related
to SymbolicTile, but not changing ConstraintExpression).

PiperOrigin-RevId: 727821427
