+++ xla/service/gpu/model/BUILD
@@ -511,9 +511,9 @@ xla_cc_test(
 )
 
 cc_library(
-    name = "symbolic_tile",
-    srcs = ["symbolic_tile.cc"],
-    hdrs = ["symbolic_tile.h"],
+    name = "constraint_expression",
+    srcs = ["constraint_expression.cc"],
+    hdrs = ["constraint_expression.h"],
     deps = [
         ":affine_map_evaluator",
         "//xla/hlo/analysis:indexing_analysis",
@@ -522,6 +522,37 @@ cc_library(
         "@com_google_absl//absl/log",
         "@com_google_absl//absl/log:check",
         "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/types:span",
+        "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:Support",
+    ],
+)
+
+xla_cc_test(
+    name = "constraint_expression_test",
+    srcs = ["constraint_expression_test.cc"],
+    deps = [
+        ":constraint_expression",
+        "//xla/hlo/analysis:indexing_analysis",
+        "//xla/hlo/analysis:indexing_test_utils",
+        "//xla/tests:xla_internal_test_main",
+        "@com_google_absl//absl/types:span",
+        "@com_google_googletest//:gtest",
+    ],
+)
+
+cc_library(
+    name = "symbolic_tile",
+    srcs = ["symbolic_tile.cc"],
+    hdrs = ["symbolic_tile.h"],
+    deps = [
+        ":affine_map_evaluator",
+        ":constraint_expression",
+        "//xla/hlo/analysis:indexing_analysis",
+        "@com_google_absl//absl/algorithm:container",
+        "@com_google_absl//absl/log",
+        "@com_google_absl//absl/log:check",
         "@com_google_absl//absl/types:span",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:IR",
